Split server handlers into per-resource files
